+++ History.md
@@ -4,6 +4,7 @@
 ## 1.0.9 / 2015-11-27
 * Fixing jade-Output without doctype
 * Fixing issue 122: Mixin Block after Mixin Call on same Line
+* Fixing issue 123: Block append not working correct.
 
 ## 1.0.8 / 2015-11-26
 * Fixing issue 120: Terse Mode not working as expected